UI and state improvements
