+++ frontend/src/App.js
@@ -5,8 +5,8 @@ import Profile from "./pages/Profile";
 import Cart from "./pages/Cart";
 import LandingPage from "./pages/LandingPage";
 import CheckoutPage from "./pages/CheckoutPage";
-import Signup from "./components/Signup";
-import Login from "./components/Login";
+import Signup from "./pages/Signup";
+import Login from "./pages/Login";
 import Men from "./pages/Men";
 import Kids from "./pages/Kids";
 import Women from "./pages/Women";
