+++ frontend/src/components/Login.jsx
@@ -71,6 +71,7 @@ const Login = () => {
           padding: 4,
           borderRadius: 2,
           boxShadow: 3,
+          mt: 5,
         }}
       >
         <Typography variant="h4" align="center" gutterBottom>
