+++ frontend/src/components/Signup.jsx
@@ -79,6 +79,7 @@ const Signup = () => {
           padding: 4,
           borderRadius: 2,
           boxShadow: 3,
+          mt: 2,
         }}
       >
         <Typography variant="h4" align="center" gutterBottom>
