+++ frontend/src/features/slices/cartSlice.js
@@ -49,7 +49,19 @@ const cartSlice = createSlice({
         state.cart = action.payload;
       })
       .addCase(removeItemFromCart.fulfilled, (state, action) => {
-        state.cart = action.payload.cart;
+        const updatedOrderItems = state.cart.orderItems.filter(
+          (item) => item.product._id !== action.meta.arg // Filter out the removed item
+        );
+
+        // Ensure safe handling of an empty cart
+        const updatedTotalPrice = action.payload.cart?.totalPrice || 0;
+
+        // Update the state immutably
+        state.cart = {
+          ...state.cart,
+          orderItems: updatedOrderItems.length > 0 ? updatedOrderItems : [],
+          totalPrice: updatedOrderItems.length > 0 ? updatedTotalPrice : 0, // Set to 0 if the cart is empty
+        };
       });
   },
 });
