+++ frontend/src/interceptors/axiosInstance.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const axiosInstance = axios.create({
-  baseURL: "http://localhost:4000/api",
+  baseURL: process.env.REACT_APP_APIURL,
 });
 
 axiosInstance.interceptors.request.use(
