+++ frontend/src/pages/Login.jsx
@@ -1,11 +1,15 @@
 import React from 'react'
 import LoginForm from '../components/Login'
+import Header from '../components/Header'
+import Footer from '../components/Footer'
 
 function Login() {
   return (
-    
+    <>
+    <Header />
     <LoginForm />
-       
+    <Footer />
+    </>
   )
 }
