+++ frontend/src/pages/Signup.jsx
@@ -1,11 +1,15 @@
 import React from 'react'
 import SignupForm from "../components/Signup"
+import Header from '../components/Header'
+import Footer from '../components/Footer'
 
 function Signup() {
   return (
-   
+   <>
+   <Header />
     <SignupForm/>
-   
+   <Footer />
+   </>
   )
 }
